app optimized for tablet - fragment_see_all_recipes_shimmer.xml has been optimized for tablet. - moved repository interfaces to domain layer.
