+++ app/src/main/java/com/jesse/ohunelo/data/repository/NotificationRepositoryImpl.kt
@@ -8,6 +8,7 @@ import com.jesse.ohunelo.data.network.data_source.RecipeNetworkDataSource
 import com.jesse.ohunelo.data.network.models.OhuneloResult
 import com.jesse.ohunelo.di.DefaultDispatcher
 import com.jesse.ohunelo.di.IODispatcher
+import com.jesse.ohunelo.domain.repository.NotificationRepository
 import com.jesse.ohunelo.util.NotificationType
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.flow.Flow
