+++ app/src/main/java/com/jesse/ohunelo/presentation/viewmodels/ResetPasswordViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.jesse.ohunelo.R
 import com.jesse.ohunelo.data.network.models.OhuneloResult
-import com.jesse.ohunelo.data.repository.AuthenticationRepository
+import com.jesse.ohunelo.domain.repository.AuthenticationRepository
 import com.jesse.ohunelo.domain.usecase.ValidateEmailUseCase
 import com.jesse.ohunelo.presentation.uistates.ResetPasswordUiState
 import com.jesse.ohunelo.util.AuthenticationException
@@ -93,7 +93,7 @@ class ResetPasswordViewModel @Inject constructor(
                             is AuthenticationException.UserTokenExpiredException -> UiText.StringResource(resId = R.string.user_token_expired)
                             is AuthenticationException.InvalidUserTokenException -> UiText.StringResource(resId = R.string.invalid_user_token)
                             is NetworkErrorException -> UiText.StringResource(resId = R.string.network_error_occured)
-                            is Exception -> UiText.StringResource(resId = R.string.login_failed)
+                            is Exception -> UiText.StringResource(resId = R.string.reset_password_email_failed)
                             else -> null
                         }
                         _resetPasswordUiStateFlow.update {
