+++ app/src/main/java/com/jesse/ohunelo/presentation/viewmodels/VerifyEmailViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.jesse.ohunelo.R
 import com.jesse.ohunelo.data.network.models.OhuneloResult
-import com.jesse.ohunelo.data.repository.AuthenticationRepository
+import com.jesse.ohunelo.domain.repository.AuthenticationRepository
 import com.jesse.ohunelo.presentation.uistates.VerifyEmailUiState
 import com.jesse.ohunelo.util.AuthenticationException
 import com.jesse.ohunelo.util.NetworkErrorException
@@ -102,7 +102,6 @@ class VerifyEmailViewModel @Inject constructor(
                     uiActionFlow.emit(UiAction.NavigateToNextScreen)
                     cancel()
                 }
-
             }
         }
     }
