+++ app/src/main/res/navigation/main_nav_graph.xml
@@ -6,6 +6,7 @@
     app:startDestination="@+id/onboardingFragment"
     >
 
+
     <fragment
         android:id="@+id/homeFragment"
         android:name="com.jesse.ohunelo.presentation.ui.fragment.HomeFragment"
